+++ examples/svelte/large-form/src/runes/form.ts
@@ -1,0 +1,12 @@
+import { createFormCreator } from '@tanstack/svelte-form'
+import TextField from '../components/text-field.svelte'
+import SubscribeButton from '../components/subscribe-button.svelte'
+
+export const { createAppForm } = createFormCreator({
+  fieldComponents: {
+    TextField,
+  },
+  formComponents: {
+    SubscribeButton,
+  },
+})