+++ packages/form-core/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @tanstack/form-core
 
+## 1.27.0
+
+### Patch Changes
+
+- Fixed issues with React Compiler ([#1893](https://github.com/TanStack/form/pull/1893))
+
+- Fix issue with deleteField and numeric keys ([#1891](https://github.com/TanStack/form/pull/1891))
+
+## 1.26.0
+
+### Patch Changes
+
+- fix stale fields on array changes ([#1729](https://github.com/TanStack/form/pull/1729))
+- allow explicitly setting `field.handleChange(undefined)` ([#1729](https://github.com/TanStack/form/pull/1729))
+
 ## 1.25.0
 
 ### Patch Changes