+++ packages/solid-form-devtools/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @tanstack/solid-form-devtools
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/form-devtools@0.2.3
+
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/form-devtools@0.2.2
+
 ## 0.2.1
 
 ### Patch Changes